requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
APScheduler==3.9.1
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
cachetools==5.0.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
frozenlist==1.3.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
motor==3.0.0
multidict==6.0.2
pymongo==4.1.1
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rfc3986==1.5.0
singleton-decorator==1.0.0
six==1.16.0
sniffio==1.2.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
yarl==1.7.2
